+++ application/controllers/Mtce.php
@@ -1,32 +1,69 @@
 <?php
 /**
  * Maintenance controller displays pages under the /mtce target link.
- * Displays all tasks with statuses using view fragments.
+ * Displays all tasks with statuses in a pagination format using view fragments.
  */
 class Mtce extends Application {
-        /**
-         * Displays the Maintenance landing page.
-         * Uses the itemlist.php and oneitem.php view fragements to display all tasks.
-         */
+
+        private $items_per_page = 10;
+
+        // displays the maintenance landing page
         public function index()
         {
-                $this->data['pagetitle'] = 'TODO List Maintenance';
-                $tasks = $this->tasks->all(); // get all the tasks
+                $this->page(1);
+        }
 
-                // substitute the status name
+        
+        private function show_page($tasks)
+        {
+                $this->data['pagetitle'] = 'TODO List Maintenance';
+                // build the task presentation output
+                $result = ''; // start with an empty array
+
                 foreach ($tasks as $task)
+                {
                         if (!empty($task->status))
                                 $task->status = $this->app->status($task->status);
-
-                // build the task presentation output
-                $result = '';   // start with an empty array        
-                foreach ($tasks as $task)
-                        $result .= $this->parser->parse('oneitem',(array)$task,true);
+                        $result .= $this->parser->parse('oneitem', (array) $task, true);
+                }
+                $this->data['display_tasks'] = $result;
 
                 // and then pass them on
-                $this->data['display_tasks'] = $result;
                 $this->data['pagebody'] = 'itemlist';
                 $this->render();
         }
 
+        // Extract & handle a page of items, defaulting to the beginning
+        public function page($num = 1)
+        {
+                $records = $this->tasks->all(); // get all the tasks
+                $tasks = array(); // start with an empty extract
+
+                // use a foreach loop, because the record indices may not be sequential
+                $index = 0; // where are we in the tasks list
+                $count = 0; // how many items have we added to the extract
+                $start = ($num - 1) * $this->items_per_page;
+                foreach($records as $task) {
+                        if ($index++ >= $start) 
+                        {
+                                $tasks[] = $task;
+                                $count++;
+                        }
+                        if ($count >= $this->items_per_page) break;
+                }
+                $this->data['pagination'] = $this->pagenav($num);
+                $this->show_page($tasks);
+        }
+
+        // Build the pagination navbar
+        private function pagenav($num) {
+                $lastpage = ceil($this->tasks->size() / $this->items_per_page);
+                $parms = array(
+                        'first' => 1,
+                        'previous' => (max($num-1,1)),
+                        'next' => min($num+1,$lastpage),
+                        'last' => $lastpage
+                );
+                return $this->parser->parse('itemnav',$parms,true);
+        }
 }