+++ changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2017-10-19
+### Added
+- itemnav.php view fragment for navigation on maintenance index page
+
+### Changed
+- itemlist view fragment to use itemnav view fragment
+- Mtce has new functions implemented to use pagination format
+
 ## [1.1.0] - 2017-10-19
 ### Added
 - Maintenance controller target link in config.php